refactor: move http client calls & vc4 entities
